.github/scripts/requirements.txt
anthropic==0.45.2
PyGithub==2.6.0
python-dotenv==1.0.0
